Fix feedback for sending in attestering
Also moved utils for behandling into its own folder
